backend/gpt/utils.py
import os
from dotenv import load_dotenv
import google.generativeai as genai

# Load environment variables from .env file
load_dotenv()

# Get the Gemini API key from environment variables
api_key = os.getenv("GEMINI_API_KEY")
if not api_key:
    raise ValueError("GEMINI_API_KEY not set in environment variables")

# Initialize Gemini client
genai.configure(api_key=api_key)

# Default chat parameters for Gemini model
GEMINI_PARAMS = dict(
    temperature=0.7,
    top_p=0.95,
)

# Model selection
MODEL = "gemini-1.5-flash"

def retrieve_documents(query):
    """
    Dummy document retrieval function.
    Replace this with your actual retrieval logic.
    Returns a list of document strings related to the query.
    """
    return [f"Document content related to query: {query}"]

def generate_response(query, context):
    """
    Generate a response using Gemini's generate content API
    based on the user query and retrieved context.
    
    Args:
        query (str): User's question.
        context (str): Retrieved document/context relevant to the query.
    
    Returns:
        str: Generated answer from the model.
    """
    prompt = (
        f"Answer the question based on the following context:\n"
        f"Context: {context}\n"
        f"Question: {query}"
    )
    
    model = genai.GenerativeModel(MODEL)
    response = model.generate_content(prompt, generation_config=genai.GenerationConfig(**GEMINI_PARAMS))
    
    return response.text

def process_file(file_url):
    """
    Placeholder function to process a file given by URL.
    Replace this with your real file fetching and text extraction logic.
    
    Args:
        file_url (str): URL or path to the file.
        
    Returns:
        str: Extracted text content from the file.
    """
    print(f"Processing file: {file_url}")
    # TODO: Add real file download and text extraction here
    return f"Extracted text content from {file_url}"

# Example usage (remove or comment out in production)
if __name__ == "__main__":
    query = "What is Retrieval-Augmented Generation?"
    docs = retrieve_documents(query)
    combined_context = " ".join(docs)
    answer = generate_response(query, combined_context)
    print("Generated Answer:", answer)
    
    file_content = process_file("https://example.com/sample.pdf")
    print("File Content:", file_content)